Refactor CLI into package and improve defaults
