--- doc_ai/cli.py
@@ -1,489 +1,0 @@
-#!/usr/bin/env python3
-"""CLI orchestrator for AI document analysis pipeline."""
-from __future__ import annotations
-
-from pathlib import Path
-from typing import List, Optional
-import os
-import sys
-import shlex
-import re
-import traceback
-import warnings
-import logging
-
-import typer
-from rich.console import Console
-from dotenv import load_dotenv
-
-# Ensure project root is first on sys.path when running as a script.
-if __package__ in (None, ""):
-    sys.path[0] = str(Path(__file__).resolve().parent.parent)
-
-from doc_ai.converter import OutputFormat, convert_path, suffix_for_format
-from doc_ai.github import build_vector_store, run_prompt, validate_file
-from doc_ai.metadata import (
-    compute_hash,
-    is_step_done,
-    load_metadata,
-    mark_step,
-    save_metadata,
-)
-
-load_dotenv()
-
-console = Console()
-app = typer.Typer(
-    help="Orchestrate conversion, validation, analysis and embedding generation."
-)
-
-SETTINGS = {"verbose": os.getenv("VERBOSE", "").lower() in {"1", "true", "yes"}}
-
-@app.callback()
-def main(verbose: bool = typer.Option(False, "--verbose", "-v", help="Enable verbose output")):
-    """Global options."""
-    SETTINGS["verbose"] = verbose
-
-ASCII_ART = r"""
- ____   ___   ____      _    ___    ____ _     ___
-|  _ \ / _ \ / ___|    / \  |_ _|  / ___| |   |_ _|
-| | | | | | | |       / _ \  | |  | |   | |    | |
-| |_| | |_| | |___   / ___ \ | |  | |___| |___ | |
-|____/ \___/ \____| /_/   \_\___|  \____|_____|___|
-"""
-
-def _print_banner() -> None:  # pragma: no cover - visual flair only
-    console.print(f"[bold green]{ASCII_ART}[/bold green]")
-
-def _suffix(fmt: OutputFormat) -> str:
-    return f".converted{suffix_for_format(fmt)}"
-
-
-def _infer_format(path: Path) -> OutputFormat:
-    mapping = {
-        ".md": OutputFormat.MARKDOWN,
-        ".html": OutputFormat.HTML,
-        ".json": OutputFormat.JSON,
-        ".txt": OutputFormat.TEXT,
-        ".doctags": OutputFormat.DOCTAGS,
-        ".dogtags": OutputFormat.DOCTAGS,
-    }
-    try:
-        return mapping[path.suffix.lower()]
-    except KeyError as exc:
-        valid = ", ".join(mapping.keys())
-        raise typer.BadParameter(
-            f"Unknown file extension '{path.suffix}'. Expected one of: {valid}"
-        ) from exc
-
-
-def _parse_env_formats() -> List[OutputFormat] | None:
-    """Return formats from OUTPUT_FORMATS env var if set."""
-
-    env_val = os.getenv("OUTPUT_FORMATS")
-    if not env_val:
-        return None
-    formats: List[OutputFormat] = []
-    for val in env_val.split(","):
-        try:
-            formats.append(OutputFormat(val.strip()))
-        except ValueError as exc:
-            valid = ", ".join(f.value for f in OutputFormat)
-            raise typer.BadParameter(
-                f"Invalid output format '{val}'. Choose from: {valid}"
-            ) from exc
-    return formats
-
-
-@app.command()
-def settings(
-    verbose: bool = typer.Option(
-        None, "--verbose/--no-verbose", help="Toggle verbose error output"
-    )
-) -> None:
-    """Show or update runtime settings."""
-    if verbose is not None:
-        SETTINGS["verbose"] = verbose
-    console.print("Current settings:")
-    console.print(f"  verbose: {SETTINGS['verbose']}")
-    env_vars = [
-        "OUTPUT_FORMATS",
-        "OPENAI_API_KEY",
-        "GITHUB_TOKEN",
-        "BASE_MODEL_URL",
-        "VALIDATE_BASE_MODEL_URL",
-    ]
-    for var in env_vars:
-        console.print(f"  {var}: {os.getenv(var, '')}")
-
-
-
-def validate_doc(
-    raw: Path,
-    rendered: Path,
-    fmt: OutputFormat | None = None,
-    prompt: Path | None = None,
-    model: str | None = None,
-    base_url: str | None = None,
-    show_progress: bool = False,
-    logger: logging.Logger | None = None,
-    console: Console | None = None,
-) -> None:
-    meta = load_metadata(raw)
-    file_hash = compute_hash(raw)
-    if meta.blake2b == file_hash and is_step_done(meta, "validation"):
-        return
-    if meta.blake2b != file_hash:
-        meta.blake2b = file_hash
-        meta.extra = {}
-    if fmt is None:
-        mapping = {
-            ".md": OutputFormat.MARKDOWN,
-            ".html": OutputFormat.HTML,
-            ".json": OutputFormat.JSON,
-            ".txt": OutputFormat.TEXT,
-            ".doctags": OutputFormat.DOCTAGS,
-        }
-        fmt = mapping.get(rendered.suffix)
-        if fmt is None:
-            valid = ", ".join(mapping.keys())
-            raise typer.BadParameter(
-                f"Unknown file extension '{rendered.suffix}'. Expected one of: {valid}"
-            )
-    prompt_path = prompt
-    if prompt_path is None:
-        doc_prompt = raw.with_name(f"{raw.stem}.validate.prompt.yaml")
-        dir_prompt = raw.with_name("validate.prompt.yaml")
-        if doc_prompt.exists():
-            prompt_path = doc_prompt
-        elif dir_prompt.exists():
-            prompt_path = dir_prompt
-        else:
-            prompt_path = Path(
-                ".github/prompts/validate-output.validate.prompt.yaml"
-            )
-    verdict = validate_file(
-        raw,
-        rendered,
-        fmt,
-        prompt_path,
-        model=model,
-        base_url=base_url,
-        show_progress=show_progress,
-        logger=logger,
-        console=console,
-    )
-    if not verdict.get("match", False):
-        raise RuntimeError(f"Mismatch detected: {verdict}")
-    mark_step(
-        meta,
-        "validation",
-        outputs=[rendered.name],
-        inputs={
-            "prompt": prompt_path.name,
-            "rendered": rendered.name,
-            "format": fmt.value,
-        },
-    )
-    save_metadata(raw, meta)
-
-
-def analyze_doc(
-    markdown_doc: Path,
-    prompt: Path | None = None,
-    output: Path | None = None,
-    model: str | None = None,
-    base_url: str | None = None,
-) -> None:
-    step_name = "analysis"
-    raw_doc = markdown_doc
-    if ".converted" in markdown_doc.suffixes:
-        raw_doc = raw_doc.with_suffix("").with_suffix("")
-    meta = load_metadata(raw_doc)
-    md_hash = compute_hash(markdown_doc)
-    prev_hash = None
-    if meta.extra:
-        prev_hash = (
-            meta.extra.get("inputs", {})
-            .get(step_name, {})
-            .get("markdown_blake2b")
-        )
-    if is_step_done(meta, step_name) and prev_hash == md_hash:
-        return
-
-    prompt_path = prompt
-    if prompt_path is None:
-        type_prompt = markdown_doc.parent / (
-            f"{markdown_doc.parent.name}.analysis.prompt.yaml"
-        )
-        dir_prompt = markdown_doc.parent / "analysis.prompt.yaml"
-        if type_prompt.exists():
-            prompt_path = type_prompt
-        elif dir_prompt.exists():
-            prompt_path = dir_prompt
-        else:
-            prompt_path = Path(
-                ".github/prompts/doc-analysis.analysis.prompt.yaml"
-            )
-
-    result = run_prompt(
-        prompt_path,
-        markdown_doc.read_text(),
-        model=model,
-        base_url=base_url,
-    )
-    result = result.strip()
-    fence = re.match(r"```(?:json)?\n([\s\S]*?)\n```", result)
-    if fence:
-        result = fence.group(1).strip()
-    if output:
-        out_path = output
-    else:
-        base = markdown_doc
-        if base.suffix == ".md":
-            base = base.with_suffix("")
-        if base.suffix == ".converted":
-            base = base.with_suffix("")
-        out_path = base.with_name(f"{base.name}.analysis.json")
-    out_path.write_text(result + "\n", encoding="utf-8")
-    console.print(
-        f"[green]Analyzed {markdown_doc} -> {out_path} (SUCCESS)[/]"
-    )
-    mark_step(
-        meta,
-        step_name,
-        outputs=[out_path.name],
-        inputs={
-            "prompt": prompt_path.name,
-            "markdown": markdown_doc.name,
-            "markdown_blake2b": md_hash,
-        },
-    )
-    save_metadata(raw_doc, meta)
-
-
-@app.command()
-def convert(
-    source: str = typer.Argument(
-        ..., help="Path or URL to raw document or folder"
-    ),
-    format: List[OutputFormat] = typer.Option(
-        None,
-        "--format",
-        "-f",
-        help="Desired output format(s). Can be passed multiple times.",
-    ),
-) -> None:
-    """Convert files using Docling."""
-    env_fmts = _parse_env_formats()
-    fmts = format or env_fmts or [OutputFormat.MARKDOWN]
-    if not SETTINGS["verbose"]:
-        warnings.filterwarnings("ignore")
-    if source.startswith(("http://", "https://")):
-        results = convert_path(source, fmts)
-    else:
-        results = convert_path(Path(source), fmts)
-    if not results:
-        console.print("No new files to process.")
-
-
-@app.command()
-def validate(
-    raw: Path = typer.Argument(..., help="Path to raw document"),
-    rendered: Path | None = typer.Argument(
-        None, help="Path to converted file"),
-    fmt: Optional[OutputFormat] = typer.Option(None, "--format", "-f"),
-    prompt: Optional[Path] = typer.Option(
-        None,
-        "--prompt",
-        help="Prompt file (overrides auto-detected *.validate.prompt.yaml)",
-    ),
-    model: Optional[str] = typer.Option(
-        None, "--model", help="Model name override"
-    ),
-    base_model_url: Optional[str] = typer.Option(
-        None, "--base-model-url", help="Model base URL override"
-    ),
-    verbose: bool = typer.Option(False, "--verbose", "-v", help="Enable debug logging"),
-    log_file: Optional[Path] = typer.Option(
-        None, "--log-file", help="Write request/response details to this file"
-    ),
-) -> None:
-    """Validate converted output against the original file."""
-    logger: logging.Logger | None = None
-    log_path = log_file
-    console = Console()
-    if verbose or log_path is not None:
-        logger = logging.getLogger("doc_ai.validate")
-        logger.setLevel(logging.DEBUG)
-        if verbose:
-            from rich.logging import RichHandler
-
-            sh = RichHandler(console=console, show_time=False)
-            sh.setLevel(logging.DEBUG)
-            logger.addHandler(sh)
-        if log_path is None:
-            log_path = raw.with_suffix(".validate.log")
-        log_path.parent.mkdir(parents=True, exist_ok=True)
-        fh = logging.FileHandler(log_path)
-        fh.setLevel(logging.DEBUG)
-        logger.addHandler(fh)
-
-    if rendered is None:
-        used_fmt = fmt or OutputFormat.MARKDOWN
-        rendered = raw.with_name(raw.name + _suffix(used_fmt))
-    else:
-        used_fmt = fmt or _infer_format(rendered)
-
-    validate_doc(
-        raw,
-        rendered,
-        used_fmt,
-        prompt,
-        model,
-        base_model_url,
-        show_progress=True,
-        logger=logger,
-        console=console,
-    )
-
-
-@app.command()
-def analyze(
-    source: Path = typer.Argument(..., help="Raw or converted document"),
-    fmt: Optional[OutputFormat] = typer.Option(
-        None, "--format", "-f", help="Format of converted file"
-    ),
-    prompt: Path | None = typer.Option(
-        None,
-        "--prompt",
-        "-p",
-        help="Prompt file (overrides auto-detected *.analysis.prompt.yaml)",
-    ),
-    output: Optional[Path] = typer.Option(
-        None,
-        "--output",
-        help="Optional output file; defaults to <doc>.analysis.json",
-    ),
-    model: Optional[str] = typer.Option(
-        None, "--model", help="Model name override"
-    ),
-    base_model_url: Optional[str] = typer.Option(
-        None, "--base-model-url", help="Model base URL override"
-    ),
-) -> None:
-    """Run an analysis prompt against a converted document."""
-    markdown_doc = source
-    if ".converted" not in "".join(markdown_doc.suffixes):
-        used_fmt = fmt or OutputFormat.MARKDOWN
-        markdown_doc = source.with_name(source.name + _suffix(used_fmt))
-    analyze_doc(markdown_doc, prompt, output, model, base_model_url)
-
-
-@app.command()
-def embed(
-    source: Path = typer.Argument(..., help="Directory containing Markdown files"),
-) -> None:
-    """Generate embeddings for Markdown files."""
-    build_vector_store(source)
-
-
-@app.command("pipeline")
-def pipeline(
-    source: Path = typer.Argument(..., help="Directory with raw documents"),
-    prompt: Path = typer.Option(
-        Path(".github/prompts/doc-analysis.analysis.prompt.yaml"),
-        help="Analysis prompt file",
-    ),
-    format: List[OutputFormat] = typer.Option(
-        None,
-        "--format",
-        "-f",
-        help="Desired output format(s) for conversion",
-    ),
-    model: Optional[str] = typer.Option(
-        None, "--model", help="Model name override"
-    ),
-    base_model_url: Optional[str] = typer.Option(
-        None, "--base-model-url", help="Model base URL override"
-    ),
-) -> None:
-    """Run the full pipeline: convert, validate, analyze, and embed."""
-    env_fmts = _parse_env_formats()
-    fmts = format or env_fmts or [OutputFormat.MARKDOWN]
-    convert_path(source, fmts)
-    validation_prompt = Path(
-        ".github/prompts/validate-output.validate.prompt.yaml"
-    )
-    for raw_file in source.rglob("*"):
-        if not raw_file.is_file():
-            continue
-        md_file = raw_file.with_name(raw_file.name + _suffix(OutputFormat.MARKDOWN))
-        if md_file.exists():
-            validate_doc(
-                raw_file,
-                md_file,
-                OutputFormat.MARKDOWN,
-                validation_prompt,
-                model,
-                base_model_url,
-            )
-            analyze_doc(md_file, prompt=prompt, model=model, base_url=base_model_url)
-    build_vector_store(source)
-
-
-__all__ = ["app"]
-
-
-def _interactive_shell() -> None:  # pragma: no cover - CLI utility
-    try:
-        _print_banner()
-        app(prog_name="cli.py", args=["--help"])
-    except SystemExit:
-        pass
-    while True:
-        try:
-            command = input("> ").strip()
-        except (EOFError, KeyboardInterrupt):
-            break
-        if not command:
-            continue
-        if command.lower() in {"exit", "quit"}:
-            break
-        if command.startswith("cd"):
-            parts = command.split(maxsplit=1)
-            target = Path(parts[1]).expanduser() if len(parts) > 1 else Path.home()
-            try:
-                os.chdir(target)
-            except OSError as exc:
-                console.print(f"[red]{exc}[/red]")
-            continue
-        full_cmd = command
-        if SETTINGS["verbose"]:
-            full_cmd += " --verbose"
-        try:
-            app(prog_name="cli.py", args=shlex.split(full_cmd))
-        except SystemExit:
-            pass
-        except Exception as exc:  # pragma: no cover - runtime error display
-            if SETTINGS["verbose"]:
-                traceback.print_exc()
-            else:
-                console.print(f"[red]{exc}[/red]")
-
-
-if __name__ == "__main__":
-    if len(sys.argv) > 1:
-        _print_banner()
-        args = sys.argv[1:]
-        if SETTINGS["verbose"] and "--verbose" not in args and "-v" not in args:
-            args.append("--verbose")
-        try:
-            app(prog_name="cli.py", args=args)
-        except Exception as exc:  # pragma: no cover - runtime error display
-            if SETTINGS["verbose"]:
-                traceback.print_exc()
-            else:
-                console.print(f"[red]{exc}[/red]")
-    else:
-        _interactive_shell()
